Rename netcdf module
  - Rename erddapy.core.netcdf_handling to erddapy.core.netcdf
  - Refactor imports
