+++ erddapy/erddapy.py
@@ -10,7 +10,7 @@ import pandas as pd
 import pytz
 from pandas._libs.tslibs.parsing import parse_time_string
 
-from erddapy.core.netcdf_handling import _nc_dataset, _tempnc
+from erddapy.core.netcdf import _nc_dataset, _tempnc
 from erddapy.core.url import _distinct, urlopen
 from erddapy.servers.servers import servers
